# docs/evidence/v0.2a1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vehicle Photo QC",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sequence": {
      "get": {
        "summary": "Sequence",
        "operationId": "sequence_api_sequence_get",
        "parameters": [
          {
            "name": "dealership_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Dealership Id"
            }
          },
          {
            "name": "inventory_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "used",
              "title": "Inventory Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shot-types": {
      "post": {
        "summary": "Add Shot Type",
        "operationId": "add_shot_type_api_shot_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/paste": {
      "post": {
        "summary": "Paste",
        "operationId": "paste_api_training_paste_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shoots/{shoot_id}/trash": {
      "post": {
        "summary": "Trash Shoot",
        "operationId": "trash_shoot_api_shoots__shoot_id__trash_post",
        "parameters": [
          {
            "name": "shoot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shoot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrashInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}/trash": {
      "post": {
        "summary": "Trash Photo",
        "operationId": "trash_photo_api_photos__photo_id__trash_post",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrashInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trash/{kind}/{item_id}/restore": {
      "post": {
        "summary": "Restore",
        "operationId": "restore_api_trash__kind___item_id__restore_post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrashInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Labels": {
        "properties": {
          "shot_type": {
            "type": "string",
            "title": "Shot Type",
            "default": "unknown"
          },
          "angle": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "too_front",
              "too_rear",
              "other_bad"
            ],
            "title": "Angle",
            "default": "unknown"
          },
          "blur": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "bad"
            ],
            "title": "Blur",
            "default": "unknown"
          },
          "crop": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "bad"
            ],
            "title": "Crop",
            "default": "unknown"
          },
          "exposure": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "bad"
            ],
            "title": "Exposure",
            "default": "unknown"
          },
          "saturation": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "bad"
            ],
            "title": "Saturation",
            "default": "unknown"
          },
          "framing": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "bad"
            ],
            "title": "Framing",
            "default": "unknown"
          },
          "overall": {
            "type": "string",
            "enum": [
              "unknown",
              "good",
              "borderline",
              "bad"
            ],
            "title": "Overall",
            "default": "unknown"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Labels"
      },
      "NameInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NameInput"
      },
      "PasteInput": {
        "properties": {
          "targets": {
            "items": {
              "$ref": "#/components/schemas/PasteTarget"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Targets"
          },
          "labels": {
            "$ref": "#/components/schemas/Labels"
          },
          "actor": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Actor",
            "default": "Local reviewer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "targets",
          "labels"
        ],
        "title": "PasteInput"
      },
      "PasteTarget": {
        "properties": {
          "photo_id": {
            "type": "string",
            "title": "Photo Id"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "photo_id"
        ],
        "title": "PasteTarget"
      },
...
      "TrashInput": {
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "all",
              "training"
            ],
            "title": "Scope",
            "default": "all"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TrashInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}